requirements.txt
anyio==3.6.2
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.2
awscli==1.18.69
backports.zoneinfo==0.2.1
beautifulsoup4==4.7.1
blinker==1.6.2
botocore==1.16.19
Brlapi==0.7.0
bs4==0.0.1
cachecontrol==0.13.1
cachetools==5.3.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.1.3
colorama==0.4.1
command-not-found==0.3
cryptography==40.0.1
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.7
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.4.2
docutils==0.16
entrypoints==0.3
fastapi==0.95.0
filelock==3.12.2
firebase-admin==6.2.0
flask==2.3.3
Flask-Cors==4.0.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
hiredis==2.2.2
html5lib==1.0.1
httplib2==0.22.0
httptools==0.1.1
idna==2.7
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.9.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==2.1.3
more-itertools==8.14.0
msgpack==1.0.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
passlib==1.7.4
pexpect==4.6.0
phonenumbers==8.10.2
Pillow==7.0.0
platformdirs==3.10.0
pptree==3.1
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.10.7
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==2.8.0
pymacaroons==0.13.0
pymongo==4.5.0
PyNaCl==1.3.0
pyparsing==3.1.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-ulid==1.1.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==4.5.3
redis-om==0.1.2
reportlab==3.5.34
requests==2.20.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.26.1
systemd-python==234
types-pyOpenSSL==23.1.0.1
types-redis==4.5.3.1
typing-extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.24.2
usb-creator==0.3.7
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.24.3
wadllib==1.3.3
watchfiles==0.19.0
webencodings==0.5.1
websockets==10.4
werkzeug==2.3.7
wsproto==0.15.0
xkit==0.0.0
zipp==3.16.2
